Refactor field and comparison enums
