+++ iceaxe/__init__.py
@@ -1,9 +1,9 @@
 from .base import (
-    Field as Field,
     IndexConstraint as IndexConstraint,
     TableBase as TableBase,
     UniqueConstraint as UniqueConstraint,
 )
+from .field import Field as Field
 from .functions import func as func
 from .postgres import PostgresDateTime as PostgresDateTime, PostgresTime as PostgresTime
 from .queries import QueryBuilder as QueryBuilder, select as select, update as update
