+++ iceaxe/comparison.py
@@ -1,0 +1,48 @@
+from abc import ABC, abstractmethod
+from enum import StrEnum
+from typing import Any
+
+
+class ComparisonType(StrEnum):
+    EQ = "="
+    NE = "!="
+    LT = "<"
+    LE = "<="
+    GT = ">"
+    GE = ">="
+    IN = "IN"
+    NOT_IN = "NOT IN"
+    LIKE = "LIKE"
+
+
+class ComparisonBase(ABC):
+    def __eq__(self, other):  # type: ignore
+        return self._compare(ComparisonType.EQ, other)
+
+    def __ne__(self, other):  # type: ignore
+        return self._compare(ComparisonType.NE, other)
+
+    def __lt__(self, other):
+        return self._compare(ComparisonType.LT, other)
+
+    def __le__(self, other):
+        return self._compare(ComparisonType.LE, other)
+
+    def __gt__(self, other):
+        return self._compare(ComparisonType.GT, other)
+
+    def __ge__(self, other):
+        return self._compare(ComparisonType.GE, other)
+
+    def in_(self, other) -> bool:
+        return self._compare(ComparisonType.IN, other)  # type: ignore
+
+    def not_in(self, other) -> bool:
+        return self._compare(ComparisonType.NOT_IN, other)  # type: ignore
+
+    def like(self, other) -> bool:
+        return self._compare(ComparisonType.LIKE, other)  # type: ignore
+
+    @abstractmethod
+    def _compare(self, comparison: ComparisonType, other: Any) -> Any:
+        pass
